+++ internal/controller/core/connect.go
@@ -268,11 +268,13 @@ func (r *ConnectReconciler) ensureDeployedService(ctx context.Context, req ctrl.
 		}
 	}
 
-	if c.Spec.RegisterOnFirstLogin && c.Spec.Auth.Type == "" {
-		l.Info("registerOnFirstLogin is set but no auth type is configured; this setting only applies to OAuth2/OIDC and will be ignored")
-	} else if c.Spec.RegisterOnFirstLogin && c.Spec.Auth.Type != positcov1beta1.AuthTypeOidc {
-		l.Info("registerOnFirstLogin is set but auth type is not oidc; this setting only applies to OAuth2/OIDC and will be ignored",
-			"authType", c.Spec.Auth.Type)
+	if regOnFirst := c.Spec.RegisterOnFirstLogin; regOnFirst != nil && *regOnFirst {
+		if c.Spec.Auth.Type == "" {
+			l.Info("registerOnFirstLogin is set but no auth type is configured; this setting only applies to OAuth2/OIDC and will be ignored")
+		} else if c.Spec.Auth.Type != positcov1beta1.AuthTypeOidc {
+			l.Info("registerOnFirstLogin is set but auth type is not oidc; this setting only applies to OAuth2/OIDC and will be ignored",
+				"authType", c.Spec.Auth.Type)
+		}
 	}
 
 	if c.Spec.Auth.Type == positcov1beta1.AuthTypeOidc {